+++ ChangeLog.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## Version [4.3.0](https://github.com/CedX/Html.ps1/compare/v4.2.0...v4.3.0)
-- Added the `New-HtmlQuery` cmdlet.
+- Added the `New-HtmlQueryString` cmdlet.
 
 ## Version [4.2.0](https://github.com/CedX/Html.ps1/compare/v4.1.0...v4.2.0)
 - Added the `New-HtmlDataUri` cmdlet.